% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.code.set}
\alias{as.code.set}
\title{Convert codeset to contingency table}
\usage{
as.code.set(x)
}
\arguments{
\item{x}{matrix contingency table (2x2)}
}
\value{
2-column matrix representation of the contingency table
}
\description{
Convert codeset to contingency table
}
