% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.contingency.table}
\alias{as.contingency.table}
\title{Convert a codeset to a contingency table}
\usage{
as.contingency.table(x)
}
\arguments{
\item{x}{codeset}
}
\value{
contingency table as a 2x2 matrix
}
\description{
Convert a codeset to a contingency table
}
