% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baserateCT.R
\name{baserateCT}
\alias{baserateCT}
\title{Calculate Baserate (CT)}
\usage{
baserateCT(CT)
}
\arguments{
\item{CT}{The \code{\link{contingencyTable}} for which the baserate is calculated}
}
\value{
A list of the format:\describe{
\item{firstBaserate}{The percentage of the data for which a positive code, or a 1, appears in the first rater}
\item{secondBaserate}{The percentage of the data for which a positive code, or a 1, appears in the second rater}
\item{averageBaserate}{The average of the firstBaserate and secondBaserate.}
}
}
\description{
This function calculates the baserate of the first rater, second rater, and the average of both the raters. Called by \code{\link{baserate}}.
}
\seealso{
\code{\link{baserate}} and \code{\link{baserateSet}}
}
