% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baserateSet.R
\name{baserateSet}
\alias{baserateSet}
\title{Calculate Baserate (Set)}
\usage{
baserateSet(set)
}
\arguments{
\item{set}{The \code{\link{codeSet}} for which the baserate is calculated}
}
\value{
A list of the format:\describe{
\item{firstBaserate}{The percentage that a positive code, or a 1, appears in the first rater}
\item{secondBaserate}{The percentage that a positive code, or a 1, appears in the second rater}
\item{averageBaserate}{The average percentage that a positive code, or a 1, appears in either of the two raters}
}
}
\description{
This function will calculate the baserate of the first rater, second rater, and the average of both the raters. Called by \code{\link{baserate}}.
}
\seealso{
\code{\link{baserate}} and \code{\link{baserateCT}}
}
