% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{$.rating.set}
\alias{$.rating.set}
\title{Helper function to return special values on a rating set}
\usage{
\method{$}{rating.set}(x, i)
}
\arguments{
\item{x}{Set or Contingency.Table}

\item{i}{Value to search for}
}
\description{
Helper function to return special values on a rating set
}
