% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{contingency_table}
\alias{contingency_table}
\title{contingency_table}
\usage{
contingency_table(precision, rec, length, baserate)
}
\arguments{
\item{precision}{double}

\item{rec}{double}

\item{length}{int}

\item{baserate}{double}
}
\description{
Create a contingency table using the provied precision, recall, baserate, and length.
}
