% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getBootPvalue_c}
\alias{getBootPvalue_c}
\title{getBootPvalue_c}
\usage{
getBootPvalue_c(distribution, result)
}
\arguments{
\item{distribution}{vector of calculated kappas}

\item{result}{double calculated kappa to compare against}
}
\value{
double calculated p-value
}
\description{
returns the percentage of the time that the distribution was greater or equal to the observed kappa
if the result is less than the mean of the distribution, than the p value is 1
else return the number of times that the distribution is greater than the result as a percentage of the total number 
of items in the distribution
}
