% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHandSet.R
\name{getHandCT}
\alias{getHandCT}
\title{Get Handset}
\usage{
getHandCT(full.ct, handSetLength, handSetBaserate, as_kappa = TRUE)
}
\arguments{
\item{full.ct}{This is the set to take a handset of}

\item{handSetLength}{This is the length of the handset to take}

\item{handSetBaserate}{This is the minimum baserate to inflate the handset to}

\item{as_kappa}{If FALSE then return the handSet, if TRUE (default) return the kappa of the handSet}
}
\value{
The function returns the handSet if returnSet is TRUE or the kappa of the handSet if not
}
\description{
This function is to get a handset of a set and calculate the kappa
}
