% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHandSet.R
\name{getHandSet}
\alias{getHandSet}
\title{Get Handset}
\usage{
getHandSet(set, handSetLength, handSetBaserate, returnSet = FALSE)
}
\arguments{
\item{set}{This is the set to take a handset of}

\item{handSetLength}{This is the length of the handset to take}

\item{handSetBaserate}{This is the minimum baserate to inflate the handset to}

\item{returnSet}{If TRUE, then return the handSet if FALSE, return the kappa of the handSet}
}
\value{
The function returns the handSet if returnSet is TRUE or the kappa of the handSet if not
}
\description{
This function is to get a handset of a set and calculate the kappa
}
