% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHandSetIndices.R
\name{getHandSetIndices}
\alias{getHandSetIndices}
\title{Generate a Handset}
\usage{
getHandSetIndices(set, handSetLength = 20, handSetBaserate = 0.2)
}
\arguments{
\item{set}{matrix of two columns}

\item{handSetLength}{number of indices to find}

\item{handSetBaserate}{number between 0 and 1 to use as a minimum number of positive indices}
}
\value{
vector of indices from set
}
\description{
Generate a vector representing indices of set, using the handSetBaserate
to determine the minimum number of indices that are positive
}
