% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa.R
\name{kappa}
\alias{kappa}
\title{Calculate kappa}
\usage{
kappa(data)
}
\arguments{
\item{data}{A \code{\link{contingencyTable}} or a \code{\link{codeSet}}}
}
\value{
The kappa of the \code{\link{contingencyTable}} or \code{\link{codeSet}}
}
\description{
This function calculates Cohen's kappa on a \code{\link{contingencyTable}} or a \code{\link{codeSet}}
}
\examples{
#Given a code set
kappa(data = codeSet)

#Given a contingency Table
kappa(data = contingencyTable)

}
\seealso{
\code{\link{kappaSet}} and \code{\link{kappaCT}}
}
