% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappaCT.R
\name{kappaCT}
\alias{kappaCT}
\title{Calculate kappa (contingency Table)}
\usage{
kappaCT(ct)
}
\arguments{
\item{ct}{A \code{\link{contingencyTable}}}
}
\value{
The kappa of the \code{\link{contingencyTable}}
}
\description{
This function calculates Cohen's kappa on a \code{\link{contingencyTable}}. Called by \code{\link{kappa}}.
}
\seealso{
\code{\link{kappa}} and \code{\link{kappaSet}}
}
