% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappaSet.R
\name{kappaSet}
\alias{kappaSet}
\title{Calculate kappa (Set)}
\usage{
kappaSet(set)
}
\arguments{
\item{set}{A \code{\link{codeSet}}}
}
\value{
The kappa of the \code{\link{codeSet}}
}
\description{
This function calculates Cohen's kappa for a given \code{\link{codeSet}}. Called by \code{\link{kappa}}.
}
\seealso{
\code{\link{kappa}} and \code{\link{kappaCT}}
}
