% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{random_contingency_table}
\alias{random_contingency_table}
\title{random_contingency_table}
\usage{
random_contingency_table(
  setLength,
  baserate,
  kappaMin,
  kappaMax,
  minPrecision = 0,
  maxPrecision = 1
)
}
\arguments{
\item{setLength}{[TBD]}

\item{baserate}{[TBD]}

\item{kappaMin}{[TBD]}

\item{kappaMax}{[TBD]}

\item{minPrecision}{[TBD]}

\item{maxPrecision}{[TBD]}
}
\description{
random_contingency_table
}
