% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{recall}
\alias{recall}
\title{recall}
\usage{
recall(kappa, BR, P)
}
\arguments{
\item{kappa}{double}

\item{BR}{double}

\item{P}{double}
}
\value{
Recall calculated from provided kappa, BR, and P
}
\description{
recall
}
