% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_contingency_table}
\alias{sample_contingency_table}
\title{sample_contingency_table}
\usage{
sample_contingency_table(xx, n, forR = TRUE)
}
\arguments{
\item{xx}{contingency table matrix}

\item{n}{int size of the contingency table}

\item{forR}{bool if true, add 1 to the results accounting for R indices starting at 1}
}
\description{
sample_contingency_table
}
